feat: add message creation binary
